# PetStore.Web/src/PetStore.Web/wwwroot/openapi-original.json
{
  "openapi" : "3.0.2",
  "info" : {
    "contact" : {
      "email" : "apiteam@swagger.io"
    },
    "description" : "This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about\nSwagger at [http://swagger.io](http://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!\nYou can now help us improve the API whether it's by making changes to the definition itself or to the code.\nThat way, with time, we can improve the API in general, and expose some of the new features in OAS3.\n\nSome useful links:\n- [The Pet Store repository](https://github.com/swagger-api/swagger-petstore)\n- [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "termsOfService" : "http://swagger.io/terms/",
    "title" : "Swagger Petstore - OpenAPI 3.0",
    "version" : "1.0.17"
  },
...
  "servers" : [ {
    "url" : "/v3"
  } ],
...
  "paths" : {
...
    "/store/inventory" : {
      "get" : {
        "description" : "Returns a map of status codes to quantities",
        "operationId" : "getInventory",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "additionalProperties" : {
                    "format" : "int32",
                    "type" : "integer"
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "successful operation"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Returns pet inventories by status",
        "tags" : [ "store" ],
        "x-swagger-router-controller" : "OrderController"
      }
    },
    "/store/order" : {
      "post" : {
        "description" : "Place a new order in the store",
        "operationId" : "placeOrder",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Order"
              }
            },
            "application/xml" : {
              "schema" : {
                "$ref" : "#/components/schemas/Order"
              }
            },
            "application/x-www-form-urlencoded" : {
              "schema" : {
                "$ref" : "#/components/schemas/Order"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Order"
                }
              }
            },
            "description" : "successful operation"
          },
          "405" : {
            "description" : "Invalid input"
          }
        },
        "summary" : "Place an order for a pet",
        "tags" : [ "store" ],
        "x-swagger-router-controller" : "OrderController"
      }
    },
    "/store/order/{orderId}" : {
      "delete" : {
        "description" : "For valid response try integer IDs with value < 1000. Anything above 1000 or nonintegers will generate API errors",
        "operationId" : "deleteOrder",
        "parameters" : [ {
          "description" : "ID of the order that needs to be deleted",
          "explode" : false,
          "in" : "path",
          "name" : "orderId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "400" : {
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "description" : "Order not found"
          }
        },
        "summary" : "Delete purchase order by ID",
        "tags" : [ "store" ],
        "x-swagger-router-controller" : "OrderController"
      },
      "get" : {
        "description" : "For valid response try integer IDs with value <= 5 or > 10. Other values will generate exceptions.",
        "operationId" : "getOrderById",
        "parameters" : [ {
          "description" : "ID of order that needs to be fetched",
          "explode" : false,
          "in" : "path",
          "name" : "orderId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Order"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Order"
                }
              }
            },
            "description" : "successful operation"
          },
          "400" : {
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "description" : "Order not found"
          }
        },
        "summary" : "Find purchase order by ID",
        "tags" : [ "store" ],
        "x-swagger-router-controller" : "OrderController"
      }
    },
...
  },
  "components" : {
...
    "schemas" : {
      "Order" : {
        "example" : {
          "petId" : 198772,
          "quantity" : 7,
          "id" : 10,
          "shipDate" : "2000-01-23T04:56:07.000+00:00",
          "complete" : true,
          "status" : "approved"
        },
        "properties" : {
          "id" : {
            "example" : 10,
            "format" : "int64",
            "type" : "integer"
          },
          "petId" : {
            "example" : 198772,
            "format" : "int64",
            "type" : "integer"
          },
          "quantity" : {
            "example" : 7,
            "format" : "int32",
            "type" : "integer"
          },
          "shipDate" : {
            "format" : "date-time",
            "type" : "string"
          },
          "status" : {
            "description" : "Order Status",
            "enum" : [ "placed", "approved", "delivered" ],
            "example" : "approved",
            "type" : "string"
          },
          "complete" : {
            "type" : "boolean"
          }
        },
        "type" : "object",
        "xml" : {
          "name" : "order"
        },
        "x-swagger-router-model" : "io.swagger.petstore.model.Order"
      },
...
    },
    "securitySchemes" : {
      "petstore_auth" : {
        "flows" : {
          "implicit" : {
            "authorizationUrl" : "https://petstore.swagger.io/oauth/authorize",
            "scopes" : {
              "write:pets" : "modify pets in your account",
              "read:pets" : "read your pets"
            }
          }
        },
        "type" : "oauth2"
      },
      "api_key" : {
        "in" : "header",
        "name" : "api_key",
        "type" : "apiKey"
      }
    }
  }
}